add error components to relations
